feat(training): Implement CNN training module stubs and fix tests
- Implement all CNN training stub functions (~1100 lines):
  - Layer construction: conv2d, pool, batch_norm, dropout, dense, flatten
  - Training API: forward, backward, step, train_epoch, validate, fit
  - Data pipeline: dataloader_create, next_batch, augment_batch
  - CNN-to-SNN conversion: converter_create, convert, finetune_stdp
  - Utility: get_output_shape

- Fix regression test type mismatch (size_t* -> uint32_t* for tensor dims)

Test results: 97 tests passing
- Unit tests: 52/52 passed
- E2E tests: 11/12 passed (1 skipped - audio FFT)
- Integration tests: 15/15 passed
- Regression tests: 19/19 passed

+++ test/regression/training/test_cnn_cortex_bridge_regression.cpp
@@ -205,7 +205,7 @@ TEST_F(CNNCortexBridgeRegressionTest, GradientBufferMemoryStability) {
     ASSERT_NE(bridge, nullptr);
 
     // Create test gradient tensor
-    size_t dims[] = {128};
+    uint32_t dims[] = {128};
     nimcp_tensor_t* grad = nimcp_tensor_create(dims, 1, NIMCP_DTYPE_F32);
     ASSERT_NE(grad, nullptr);
 
@@ -498,7 +498,7 @@ TEST_F(CNNCortexBridgeRegressionTest, DisabledGradientFeedbackNoOp) {
     ASSERT_NE(bridge, nullptr);
 
     // Create gradient tensor
-    size_t dims[] = {64};
+    uint32_t dims[] = {64};
     nimcp_tensor_t* grad = nimcp_tensor_create(dims, 1, NIMCP_DTYPE_F32);
     ASSERT_NE(grad, nullptr);
